% Generated by roxygen2 (4.0.1): do not edit by hand
\name{labels<-}
\alias{labels<-}
\alias{labels<-.dendrogram}
\title{Set the labels of an object}
\usage{
labels(x,...) <- value

labels(x,...) <- value
}
\arguments{
\item{x}{Object on which to set labels}

\item{value}{New labels}

\item{...}{Additional parameters passed to specific methods}
}
\value{
object of class dendrogram
}
\description{
Set the labels of an object

Set the labels of a dendrogram
}
\examples{
hc <- hclust(dist(USArrests), "ave")
dend <- as.dendrogram(hc)
labels(dend)<-abbreviate(labels(dend),minlength=2)
}
\author{
jefferis
}
\seealso{
\code{\link{dendrogram},\link{labels}}
}

