\docType{package}
\name{dendroextras-package}
\alias{dendroextras}
\alias{dendroextras-package}
\title{Extra functions to cut, label and colour dendrogram clusters}
\description{
  Extra functions to cut, label and colour dendrogram
  clusters
}
\seealso{
  \code{\link{dendrogram}, \link{hclust}} in
  \code{\link{stats}} package.
}
\keyword{dendrogram}
\keyword{package,}

